Fix issues with where LMDB is depended on
