+++ crates/atlaspack_core/Cargo.toml
@@ -15,8 +15,6 @@ atlaspack_sourcemap = { path = "../atlaspack_sourcemap" }
 atlaspack_filesystem = { path = "../atlaspack_filesystem" }
 atlaspack_memoization_cache = { path = "../atlaspack_memoization_cache" }
 caniuse_database = { path = "../caniuse_database" }
-lmdb-js-lite = { path = "../lmdb-js-lite" }
-
 anyhow = { workspace = true }
 async-trait = { workspace = true }
 bitflags = { workspace = true }
